// src/types/types.ts
// Generated by https://quicktype.io
export interface FilterResult {
    filter: string,
    setFilter: React.Dispatch<React.SetStateAction<string>>
}

export interface Slides {
    hero:     Hero;
    slides:   Slide[];
    topSales: TopSales;
}

export interface DB {
    data?: Slide[],
    slidesCA?: Slide[],
    slidesUK?: Slide[],
}

export interface Hero {
    image: HeroImage[];
    title: string;
}

export interface HeroImage {
    url:  string;
    text: string;
}

export interface Slide {
    name:            string;
    description:     string;
    slug:            string;
    type:            string;
    title:           string;
    icon:            string;
    background:      string;
    cards:           CardType[];
    totalCampaigns?: number;
}

export interface CardType {
    id?:               number;
    title:            string;
    destination:      string;
    images:           CardImage[];
    days:             number;
    url?:              string;
    map?:              null;
    highlights?:       null;
    includes?:         null;
    priceDetail:      PriceDetail;
    tags?:             null;
    crafterDetail?:    null;
    hasPrivateTour?:   boolean;
    hasSoloTraveller?: boolean;
    headLine?:         null;
    tagHeadLine?:      any[];
    isLandOnly?:       boolean;
    isValid?:          boolean;
}

export interface CardImage {
    desktop?: string;
    tablet?:  string;
    mobile?:  string;
}

export interface PriceDetail {
    fromPrice:         number;
    fromPriceBeautify: string;
    discountSaved:     number;
    oldPriceBeautify:  string;
    pricingPercentage: number;
    pricePerNight:     string;
}

export interface TopSales {
    type:       string;
    title:      string;
    icon:       string;
    background: string;
    cards:      CardType[];
}